fixed: Order API

+++ app/api/v1/orders/route.ts
@@ -1,3 +1,5 @@
+import { getOrder } from "@/app/actions/orders/orders";
+import { getUserById } from "@/app/actions/user/auth";
 import { verifyUserToken } from "@/app/actions/user/userToken";
 import prisma from "@/lib/db";
 import { NextRequest, NextResponse } from "next/server";
@@ -17,8 +19,9 @@ type Product = {
 };
 
 type RequestData = {
-  trxId: string;
   userId: string;
+  name: string;
+  trxId?: string;
   address: string;
   contact: string;
   products: Product[];
@@ -33,33 +36,54 @@ export const dynamic = "force-dynamic";
 export async function POST(request: NextRequest) {
   try {
     const data: RequestData = await request.json();
-    // const token = await verifyUserToken();
-    // if (!token) {
-    //   return NextResponse.json({
-    //     status: 401,
-    //     success: false,
-    //     message: "Invalid request",
-    //   });
-    // }
-    const isDuplicate = await prisma.order.findUnique({
-      where: { trxId: data.trxId },
-    });
-    if (isDuplicate) {
+
+    const isUser = await getUserById(data.userId);
+    if (!isUser) {
       return NextResponse.json({
-        status: 401,
+        status: 400,
         success: false,
-        message: "Invalid TRX ID",
+        message: "Invalid request or no user found",
       });
+    }
+
+    if (!data.products) {
+      return NextResponse.json({
+        status: 400,
+        success: false,
+        message: "Missing required 'products' property in request body",
+      });
+    }
+
+    for (const product of data.products) {
+      if (
+        !(await prisma.product.findUnique({ where: { id: product.productId } }))
+      ) {
+        return NextResponse.json({
+          status: 401,
+          success: false,
+          message: "Product not found",
+        });
+      }
+      for (const extras of product.extras) {
+        if (!(await prisma.extras.findUnique({ where: { id: extras.id } }))) {
+          return NextResponse.json({
+            status: 401,
+            success: false,
+            message: "Extras not found",
+          });
+        }
+      }
     }
 
     const order = await prisma.order.create({
       data: {
         userId: data.userId,
+        name: data.name,
         trxId: data.trxId,
-        address: "Bhara Kahu",
-        contact: "03125770904",
-        quantity: 5,
-        total: 4.56,
+        address: data.address,
+        contact: data.contact,
+        quantity: data.quantity,
+        total: data.total,
         orderType:
           data.orderType === "pickup"
             ? "pickup"
@@ -90,13 +114,16 @@ export async function POST(request: NextRequest) {
           },
         });
       });
-
       return Promise.all(extraPromises);
     });
 
     await Promise.all(productPromises);
 
-    return NextResponse.json({ status: 200, order });
+    return NextResponse.json({
+      status: 200,
+      success: true,
+      message: "Order placed",
+    });
   } catch (error) {
     console.log(error);
     return NextResponse.json({
@@ -110,6 +137,15 @@ export async function POST(request: NextRequest) {
 export async function DELETE(request: NextRequest) {
   try {
     const { orderId } = await request.json();
+
+    const isOrder = await getOrder(orderId);
+    if (!isOrder) {
+      return NextResponse.json({
+        status: 404,
+        success: false,
+        message: "No order found",
+      });
+    }
 
     // Delete OrderProductExtras
     await prisma.orderProductExtras.deleteMany({
